pre-memory bound

+++ include/reduce.h
@@ -3,7 +3,7 @@
 
 #define MAX_BLOCK_SZ 1024
 
-double gpu_sum_reduce(double* d_in, unsigned int d_in_len);
+float gpu_sum_reduce(float* d_in, unsigned int d_in_len);
 void debug();
 
 #endif // !REDUCE_H__
